+++ Fence/Fence/Views/MainWindow.axaml.cs
@@ -14,8 +14,8 @@ using System.ComponentModel;
 using System.IO;
 using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Reflection;
 using System.Threading;
+using Radoub.UI.Utils;
 using Radoub.UI.Views;
 
 namespace MerchantEditor.Views;
@@ -420,38 +420,12 @@ public partial class MainWindow : Window, INotifyPropertyChanged
         {
             ToolName = "Fence",
             Subtitle = "Merchant Editor for Neverwinter Nights",
-            Version = GetVersionString(),
+            Version = VersionHelper.GetVersion(),
             IconBitmap = new Avalonia.Media.Imaging.Bitmap(
                 Avalonia.Platform.AssetLoader.Open(
                     new System.Uri("avares://Fence/Assets/fence.ico")))
         });
         aboutWindow.Show(this);
-    }
-
-    /// <summary>
-    /// Gets the version string from assembly metadata.
-    /// </summary>
-    private static string GetVersionString()
-    {
-        try
-        {
-            var infoVersion = Assembly.GetExecutingAssembly()
-                .GetCustomAttribute<AssemblyInformationalVersionAttribute>()?.InformationalVersion;
-
-            if (!string.IsNullOrEmpty(infoVersion))
-            {
-                var plusIndex = infoVersion.IndexOf('+');
-                if (plusIndex > 0)
-                    infoVersion = infoVersion[..plusIndex];
-                return infoVersion;
-            }
-
-            var version = Assembly.GetExecutingAssembly().GetName().Version;
-            if (version != null)
-                return $"{version.Major}.{version.Minor}.{version.Build}";
-        }
-        catch { }
-        return "1.0.0";
     }
 
     #endregion